Clean code and run swiftlint
