+++ MercadoPagoSDK/MercadoPagoSDK/MercadoPagoCheckoutScreens.swift
@@ -280,132 +280,4 @@ extension MercadoPagoCheckout {
 
         self.navigationController.pushViewController(entityTypeStep, animated: true)
     }
-
-    func showHookScreen(hookStep: PXHookStep) {
-
-        if let targetHook = MercadoPagoCheckoutViewModel.flowPreference.getHookForStep(hookStep: hookStep) {
-
-            let vc = MercadoPagoUIViewController()
-            vc.view.backgroundColor = .clear
-
-            vc.callbackCancel = {
-                self.viewModel.wentBackFrom(hook: hookStep)
-            }
-
-            if self.viewModel.copyViewModelAndAssignToCheckoutStore() {
-                targetHook.didReceive?(hookStore: PXCheckoutStore.sharedInstance)
-            }
-
-            if let navTitle = targetHook.titleForNavigationBar?() {
-                vc.title = navTitle
-            }
-            
-            if let navBarColor = targetHook.colorForNavigationBar?() {
-                vc.setNavBarBackgroundColor(color: navBarColor)
-            }
-            
-            vc.shouldShowBackArrow = true
-            if let shouldShowBackArrow = targetHook.shouldShowBackArrow?() {
-                 vc.shouldShowBackArrow = shouldShowBackArrow
-            }
-            
-            if let shouldShowNavigationBar = targetHook.shouldShowNavigationBar?() {
-                vc.shouldHideNavigationBar = !shouldShowNavigationBar
-            }
-            
-            let hookView = targetHook.render()
-            hookView.removeFromSuperview()
-            hookView.frame = vc.view.frame
-            vc.view.addSubview(hookView)
-
-            targetHook.renderDidFinish?()
-
-            self.navigationController.pushViewController(vc, animated: true)
-
-            self.viewModel.continueFrom(hook: hookStep)
-        }
-    }
-
-    func showPaymentMethodPluginPaymentScreen() {
-        
-        guard let paymentMethodPlugin = self.viewModel.paymentOptionSelected as? PXPaymentMethodPlugin else {
-            return
-        }
-        
-        let containerVC = MercadoPagoUIViewController()
-
-        // By feature definition. Back is not available in make payment plugin.
-        containerVC.shouldShowBackArrow = false
-
-        let paymentPluginComponent = paymentMethodPlugin.paymentPlugin
-        
-        if self.viewModel.copyViewModelAndAssignToCheckoutStore() {
-            paymentPluginComponent.didReceive?(pluginStore: PXCheckoutStore.sharedInstance)
-        }
-        
-        if let navTitle = paymentPluginComponent.titleForNavigationBar?() {
-            containerVC.title = navTitle
-        }
-        
-        if let navBarColor = paymentPluginComponent.colorForNavigationBar?() {
-            containerVC.setNavBarBackgroundColor(color: navBarColor)
-        }
-        
-        if let shouldShowNavigationBar = paymentPluginComponent.shouldShowNavigationBar?() {
-            containerVC.shouldHideNavigationBar = !shouldShowNavigationBar
-        }
-        
-        let paymentPluginComponentView = paymentPluginComponent.render()
-        paymentPluginComponentView.removeFromSuperview()
-        paymentPluginComponentView.frame = containerVC.view.frame
-        containerVC.view.addSubview(paymentPluginComponentView)
-        
-        paymentPluginComponent.renderDidFinish?()
-        
-        self.navigationController.pushViewController(containerVC, animated: true)
-    }
-
-    func showPaymentMethodPluginConfigScreen() {
-        
-        guard let paymentMethodPlugin = self.viewModel.paymentOptionSelected as? PXPaymentMethodPlugin else {
-            return
-        }
-        
-        let containerVC = MercadoPagoUIViewController()
-
-        guard let paymentMethodConfigPluginComponent = paymentMethodPlugin.paymentMethodConfigPlugin else {
-            return
-        }
-        
-        if self.viewModel.copyViewModelAndAssignToCheckoutStore() {
-            paymentMethodConfigPluginComponent.didReceive?(pluginStore: PXCheckoutStore.sharedInstance)
-        }
-        
-        if let navTitle = paymentMethodConfigPluginComponent.titleForNavigationBar?() {
-            containerVC.title = navTitle
-        }
-        
-        if let navBarColor = paymentMethodConfigPluginComponent.colorForNavigationBar?() {
-            containerVC.setNavBarBackgroundColor(color: navBarColor)
-        }
-        
-        containerVC.shouldShowBackArrow = true
-        if let shouldShowBackArrow = paymentMethodConfigPluginComponent.shouldShowBackArrow?() {
-            containerVC.shouldShowBackArrow = shouldShowBackArrow
-        }
-        
-        if let shouldShowNavigationBar = paymentMethodConfigPluginComponent.shouldShowNavigationBar?() {
-            containerVC.shouldHideNavigationBar = !shouldShowNavigationBar
-        }
-        
-        let paymentMethodConfigPluginComponentView = paymentMethodConfigPluginComponent.render()
-
-        paymentMethodConfigPluginComponentView.removeFromSuperview()
-        paymentMethodConfigPluginComponentView.frame = containerVC.view.frame
-        containerVC.view.addSubview(paymentMethodConfigPluginComponentView)
-
-        paymentMethodConfigPluginComponent.renderDidFinish?()
-        
-        self.navigationController.pushViewController(containerVC, animated: true)
-    }
 }
