+++ MercadoPagoSDK/MercadoPagoSDK/NextStepHelper.swift
@@ -278,10 +278,6 @@ extension MercadoPagoCheckoutViewModel {
         return false
     }
 
-    func needToCreatePaymentForPlugin() -> Bool {
-        return needToCreatePayment() && self.paymentOptionSelected is PXPaymentMethodPlugin
-    }
-
     func setPaymentOptionSelected() {
         guard let paymentMethod = self.paymentData.getPaymentMethod() else {
             return
@@ -317,29 +313,5 @@ extension MercadoPagoCheckoutViewModel {
             return mpESCManager.getESC(cardId: card.getCardId()) == nil ? false : true
         }
         return false
-    }
-
-    func needToShowPaymentMethodConfigPlugin() -> Bool {
-        guard let paymentMethodPluginSelected = paymentOptionSelected as? PXPaymentMethodPlugin else {
-            return false
-        }
-
-        if  wasPaymentMethodConfigPluginShowed() {
-            return false
-        }
-
-        return paymentMethodPluginSelected.paymentMethodConfigPlugin != nil
-    }
-
-    func wasPaymentMethodConfigPluginShowed() -> Bool {
-        return paymentMethodConfigPluginShowed
-    }
-
-    func willShowPaymentMethodConfigPlugin() {
-        paymentMethodConfigPluginShowed = true
-    }
-
-    func resetPaymentMethodConfigPlugin() {
-        paymentMethodConfigPluginShowed = false
     }
 }
