+++ MercadoPagoSDK/MercadoPagoSDK/PaymentVaultViewController.swift
@@ -265,9 +265,9 @@ open class PaymentVaultViewController: MercadoPagoUIScrollViewController, UIColl
     }
 
     public func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        
+
         if isGroupSection(section: indexPath.section) {
-            
+
             let paymentSearchItemSelected = self.viewModel.getPaymentMethodOption(row: indexPath.row) as! PaymentMethodOption
             collectionView.deselectItem(at: indexPath, animated: true)
             collectionView.allowsSelection = false
@@ -331,17 +331,17 @@ open class PaymentVaultViewController: MercadoPagoUIScrollViewController, UIColl
             titleCell = cell
             return cell
         } else if isGroupSection(section: indexPath.section) {
-            
+
             let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "searchCollectionCell",
-                                                          
+
                                                           for: indexPath) as! PaymentSearchCollectionViewCell
-            
+
             if let paymentMethodToDisplay = self.viewModel.getPaymentMethodOption(row: indexPath.row) {
                 cell.fillCell(drawablePaymentOption: paymentMethodToDisplay)
             }
-            
+
             return cell
-            
+
         } else if isCouponSection(section: indexPath.section) {
             let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "CouponCell", for: indexPath)
             cell.contentView.viewWithTag(1)?.removeFromSuperview()
@@ -418,7 +418,7 @@ open class PaymentVaultViewController: MercadoPagoUIScrollViewController, UIColl
 
     func heightOfItem(indexItem: Int) -> CGFloat {
         if let paymentMethodOptionDrawable = self.viewModel.getPaymentMethodOption(row: indexItem) {
-            
+
             return PaymentSearchCollectionViewCell.totalHeight(drawablePaymentOption : paymentMethodOptionDrawable)
         }
         return 0
