+++ MercadoPagoSDKExamplesObjectiveC/MercadoPagoSDKExamplesObjectiveC/Hooks/SecondHookViewController.m
@@ -40,11 +40,11 @@
 
 
 #pragma mark - PXHookComponent optional delegates.
-- (BOOL)shouldSkipHookWithHookStore:(PXHookStore * _Nonnull)hookStore {
+- (BOOL)shouldSkipHookWithHookStore:(PXCheckoutStore * _Nonnull)hookStore {
     return NO;
 }
 
-- (void)didReceiveWithHookStore:(PXHookStore * _Nonnull)hookStore {
+- (void)didReceiveWithHookStore:(PXCheckoutStore * _Nonnull)hookStore {
     self.paymentData = [hookStore getPaymentData];
 }
