+++ MercadoPagoSDKExamplesObjectiveC/MercadoPagoSDKExamplesObjectiveC/Hooks/ThirdHookViewController.m
@@ -36,7 +36,7 @@
 
 
 #pragma mark - PXHookComponent optional delegates.
-- (BOOL)shouldSkipHookWithHookStore:(PXHookStore * _Nonnull)hookStore {
+- (BOOL)shouldSkipHookWithHookStore:(PXCheckoutStore * _Nonnull)hookStore {
     return NO;
 }
